+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.38.0 - [Unreleased]
+
+### Added
+
+- Give meaning of incorrect answer, when possible. Closes [#1111](https://github.com/fniessink/toisto/issues/1111).
+
 ## 0.37.1 - 2025-09-10
 
 ### Fixed